+++ .ruby-version
@@ -1,1 +1,1 @@
-2.3.3
+2.5.0
