+++ .travis.yml
@@ -1,23 +1,15 @@
 rvm:
-  - 1.8.7
-  - 2.1.10
-  - 2.3.3
+  - 2.5.0
+  - 2.4.3
+  - 2.3.6
+  - 2.2.9
 
 gemfile:
-  - gemfiles/Gemfile.cucumber-1.3
-  - gemfiles/Gemfile.cucumber-2.4
+  - gemfiles/Gemfile.cucumber-3.1
   - gemfiles/Gemfile.cucumber-3.0
 
 matrix:
   exclude:
-    - gemfile: gemfiles/Gemfile.cucumber-3.0
-      rvm: 2.1.10
-    - gemfile: gemfiles/Gemfile.cucumber-3.0
-      rvm: 1.8.7
-    - gemfile: gemfiles/Gemfile.cucumber-2.4
-      rvm: 1.8.7
-    - gemfile: gemfiles/Gemfile.cucumber-1.3
-      rvm: 2.3.3 # doesn't work with old RSpec
 
 install:
   # Replace default Travis CI bundler script with a version that doesn't
