+++ Gemfile
@@ -1,1 +1,1 @@
-gemfiles/Gemfile.cucumber-2.4
+gemfiles/Gemfile.cucumber-3.1
