--- gemfiles/Gemfile.cucumber-1.3
@@ -1,12 +1,0 @@
-source 'https://rubygems.org'
-
-# Runtime dependencies
-gem 'cucumber', '~> 1.3.20'
-
-# Development dependencies
-gem 'rspec', '~> 1.0'
-gem 'gemika'
-gem 'rake', '=10.0.4'
-
-# Gem under test
-gem 'cucumber_priority', :path => '..'
