--- gemfiles/Gemfile.cucumber-1.3.lock
@@ -1,39 +1,0 @@
-PATH
-  remote: ..
-  specs:
-    cucumber_priority (0.1.2)
-      cucumber
-
-GEM
-  remote: https://rubygems.org/
-  specs:
-    activesupport (2.3.18)
-    builder (3.2.2)
-    cucumber (1.3.20)
-      builder (>= 2.1.2)
-      diff-lcs (>= 1.1.3)
-      gherkin (~> 2.12)
-      multi_json (>= 1.7.5, < 2.0)
-      multi_test (>= 0.1.2)
-    diff-lcs (1.2.5)
-    gemika (0.3.2)
-    gherkin (2.12.2)
-      multi_json (~> 1.3)
-    multi_json (1.11.2)
-    multi_test (0.1.2)
-    rake (10.0.4)
-    rspec (1.3.2)
-
-PLATFORMS
-  ruby
-
-DEPENDENCIES
-  activesupport (~> 2.3.0)
-  cucumber (~> 1.3.20)
-  cucumber_priority!
-  gemika
-  rake (= 10.0.4)
-  rspec (~> 1.0)
-
-BUNDLED WITH
-   1.13.7
