--- gemfiles/Gemfile.cucumber-2.4.lock
@@ -1,53 +1,0 @@
-PATH
-  remote: ..
-  specs:
-    cucumber_priority (0.2.0)
-      cucumber
-
-GEM
-  remote: https://rubygems.org/
-  specs:
-    builder (3.2.2)
-    cucumber (2.4.0)
-      builder (>= 2.1.2)
-      cucumber-core (~> 1.5.0)
-      cucumber-wire (~> 0.0.1)
-      diff-lcs (>= 1.1.3)
-      gherkin (~> 4.0)
-      multi_json (>= 1.7.5, < 2.0)
-      multi_test (>= 0.1.2)
-    cucumber-core (1.5.0)
-      gherkin (~> 4.0)
-    cucumber-wire (0.0.1)
-    diff-lcs (1.2.5)
-    gemika (0.3.2)
-    gherkin (4.0.0)
-    multi_json (1.12.1)
-    multi_test (0.1.2)
-    rake (10.1.0)
-    rspec (3.4.0)
-      rspec-core (~> 3.4.0)
-      rspec-expectations (~> 3.4.0)
-      rspec-mocks (~> 3.4.0)
-    rspec-core (3.4.3)
-      rspec-support (~> 3.4.0)
-    rspec-expectations (3.4.0)
-      diff-lcs (>= 1.2.0, < 2.0)
-      rspec-support (~> 3.4.0)
-    rspec-mocks (3.4.1)
-      diff-lcs (>= 1.2.0, < 2.0)
-      rspec-support (~> 3.4.0)
-    rspec-support (3.4.1)
-
-PLATFORMS
-  ruby
-
-DEPENDENCIES
-  cucumber (~> 2.4.0)
-  cucumber_priority!
-  gemika
-  rake
-  rspec (~> 3.0)
-
-BUNDLED WITH
-   1.13.7
