+++ gemfiles/Gemfile.cucumber-3.0
@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
 
 # Runtime dependencies
-gem 'cucumber', '~> 3.0.0.pre.1'
+gem 'cucumber', '~> 3.0.2'
 
 # Development dependencies
 gem 'rspec', '~> 3.0'
@@ -9,4 +9,4 @@ gem 'gemika'
 gem 'rake'
 
 # Gem under test
-gem 'cucumber_priority'
+gem 'cucumber_priority', :path => '..'
