+++ gemfiles/Gemfile.cucumber-3.0.lock
@@ -1,63 +1,59 @@
+PATH
+  remote: ..
+  specs:
+    cucumber_priority (0.3.0)
+      cucumber (>= 3.0.0)
+
 GEM
   remote: https://rubygems.org/
   specs:
-    activesupport (5.0.1)
-      concurrent-ruby (~> 1.0, >= 1.0.2)
-      i18n (~> 0.7)
-      minitest (~> 5.1)
-      tzinfo (~> 1.1)
-    backports (3.6.8)
-    builder (3.2.2)
-    concurrent-ruby (1.0.3)
-    cucumber (3.0.0.pre.1)
+    backports (3.11.0)
+    builder (3.2.3)
+    cucumber (3.0.2)
       builder (>= 2.1.2)
-      cucumber-core (~> 2.0)
+      cucumber-core (~> 3.0.0)
+      cucumber-expressions (~> 4.0.3)
       cucumber-wire (~> 0.0.1)
-      diff-lcs (>= 1.1.3)
+      diff-lcs (~> 1.3)
       gherkin (~> 4.0)
       multi_json (>= 1.7.5, < 2.0)
       multi_test (>= 0.1.2)
-    cucumber-core (2.0.0)
-      backports (~> 3.6)
-      gherkin (~> 4.0)
+    cucumber-core (3.0.0)
+      backports (>= 3.8.0)
+      cucumber-tag_expressions (>= 1.0.1)
+      gherkin (>= 4.1.3)
+    cucumber-expressions (4.0.4)
+    cucumber-tag_expressions (1.1.1)
     cucumber-wire (0.0.1)
-    cucumber_priority (0.1.2)
-      activesupport
-      cucumber
-    diff-lcs (1.2.5)
+    diff-lcs (1.3)
     gemika (0.3.2)
-    gherkin (4.0.0)
-    i18n (0.7.0)
-    minitest (5.10.1)
-    multi_json (1.12.1)
+    gherkin (4.1.3)
+    multi_json (1.13.1)
     multi_test (0.1.2)
-    rake (10.4.2)
-    rspec (3.5.0)
-      rspec-core (~> 3.5.0)
-      rspec-expectations (~> 3.5.0)
-      rspec-mocks (~> 3.5.0)
-    rspec-core (3.5.4)
-      rspec-support (~> 3.5.0)
-    rspec-expectations (3.5.0)
+    rake (12.3.0)
+    rspec (3.7.0)
+      rspec-core (~> 3.7.0)
+      rspec-expectations (~> 3.7.0)
+      rspec-mocks (~> 3.7.0)
+    rspec-core (3.7.1)
+      rspec-support (~> 3.7.0)
+    rspec-expectations (3.7.0)
       diff-lcs (>= 1.2.0, < 2.0)
-      rspec-support (~> 3.5.0)
-    rspec-mocks (3.5.0)
+      rspec-support (~> 3.7.0)
+    rspec-mocks (3.7.0)
       diff-lcs (>= 1.2.0, < 2.0)
-      rspec-support (~> 3.5.0)
-    rspec-support (3.5.0)
-    thread_safe (0.3.5)
-    tzinfo (1.2.2)
-      thread_safe (~> 0.1)
+      rspec-support (~> 3.7.0)
+    rspec-support (3.7.0)
 
 PLATFORMS
   ruby
 
 DEPENDENCIES
-  cucumber (~> 3.0.0.pre.1)
-  cucumber_priority
+  cucumber (~> 3.0.2)
+  cucumber_priority!
   gemika
   rake
   rspec (~> 3.0)
 
 BUNDLED WITH
-   1.13.7
+   1.16.1
