+++ gemfiles/Gemfile.cucumber-3.1
@@ -1,7 +1,7 @@
 source 'https://rubygems.org'
 
 # Runtime dependencies
-gem 'cucumber', '~> 2.4.0'
+gem 'cucumber', '~> 3.1.0'
 
 # Development dependencies
 gem 'rspec', '~> 3.0'
