+++ gemfiles/Gemfile.cucumber-3.1.lock
@@ -1,0 +1,59 @@
+PATH
+  remote: ..
+  specs:
+    cucumber (3.1.0)
+      builder (>= 2.1.2)
+      cucumber-core (~> 3.1.0)
+      cucumber-expressions (~> 5.0.4)
+      cucumber-wire (~> 0.0.1)
+      diff-lcs (~> 1.3)
+      gherkin (~> 5.0)
+      multi_json (>= 1.7.5, < 2.0)
+      multi_test (>= 0.1.2)
+    cucumber_priority (0.3.0)
+      cucumber (>= 3.0.0)
+
+GEM
+  remote: https://rubygems.org/
+  specs:
+    backports (3.11.0)
+    builder (3.2.3)
+    cucumber-core (3.1.0)
+      backports (>= 3.8.0)
+      cucumber-tag_expressions (~> 1.1.0)
+      gherkin (>= 5.0.0)
+    cucumber-expressions (5.0.13)
+    cucumber-tag_expressions (1.1.1)
+    cucumber-wire (0.0.1)
+    diff-lcs (1.3)
+    gemika (0.3.2)
+    gherkin (5.0.0)
+    multi_json (1.13.1)
+    multi_test (0.1.2)
+    rake (10.4.2)
+    rspec (3.5.0)
+      rspec-core (~> 3.5.0)
+      rspec-expectations (~> 3.5.0)
+      rspec-mocks (~> 3.5.0)
+    rspec-core (3.5.4)
+      rspec-support (~> 3.5.0)
+    rspec-expectations (3.5.0)
+      diff-lcs (>= 1.2.0, < 2.0)
+      rspec-support (~> 3.5.0)
+    rspec-mocks (3.5.0)
+      diff-lcs (>= 1.2.0, < 2.0)
+      rspec-support (~> 3.5.0)
+    rspec-support (3.5.0)
+
+PLATFORMS
+  ruby
+
+DEPENDENCIES
+  cucumber (~> 3.1.0)
+  cucumber_priority!
+  gemika
+  rake
+  rspec (~> 3.0)
+
+BUNDLED WITH
+   1.16.1
