+++ lib/cucumber_priority.rb
@@ -1,8 +1,8 @@
 require 'cucumber'
-require 'cucumber/rb_support/rb_language'
+require 'cucumber/glue/registry_and_more'
 
 require 'cucumber_priority/util'
 require 'cucumber_priority/ambiguous_error_ext'
-require 'cucumber_priority/rb_step_definition_ext'
-require 'cucumber_priority/support_code_ext'
+require 'cucumber_priority/step_definition_ext'
+require 'cucumber_priority/assert_unambiguous_match_ext'
